refactor: Move codegen_grammar into parser generator directly
