--- crates/codegen/grammar/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "codegen_grammar"
-version.workspace = true
-rust-version.workspace = true
-edition.workspace = true
-publish = false
-
-[dependencies]
-codegen_language_definition = { workspace = true }
-indexmap = { workspace = true }
-semver = { workspace = true }
-strum_macros = { workspace = true }
-
-[lints]
-workspace = true
-
