--- crates/codegen/grammar/src/lib.rs
@@ -1,15 +1,0 @@
-mod constructor;
-mod grammar;
-mod parser_definition;
-mod precedence_parser_definition;
-mod scanner_definition;
-mod version_quality;
-mod visitor;
-
-pub use constructor::GrammarConstructorDslV2;
-pub use grammar::*;
-pub use parser_definition::*;
-pub use precedence_parser_definition::*;
-pub use scanner_definition::*;
-pub use version_quality::*;
-pub use visitor::*;
