+++ crates/codegen/runtime/generator/src/parser/grammar/precedence_parser_definition.rs
@@ -1,7 +1,7 @@
 use std::fmt::Debug;
 use std::rc::Rc;
 
-use crate::{GrammarVisitor, ParserDefinitionNode, Visitable};
+use crate::parser::grammar::{GrammarVisitor, ParserDefinitionNode, Visitable};
 
 pub trait PrecedenceParserDefinition: Debug {
     fn name(&self) -> &'static str;
